Rewrited IssueTransaction in sync way
JIRA: https://wavesplatform.atlassian.net/browse/NODE-578
