--- it/src/test/scala/com/wavesplatform/it/async/transactions/IssueTransactionSuite.scala
@@ -1,91 +1,0 @@
-package com.wavesplatform.it.transactions
-
-import com.wavesplatform.it.api.AsyncHttpApi._
-import com.wavesplatform.it.util._
-import org.scalatest.prop.TableDrivenPropertyChecks
-
-import scala.concurrent.Await
-import scala.concurrent.duration._
-
-class IssueTransactionSuite extends BaseTransactionSuite with TableDrivenPropertyChecks {
-
-  private val waitCompletion = 2.minutes
-  private val defaultQuantity = 100000
-  private val assetFee = 5.waves
-
-  test("asset issue changes issuer's asset balance; issuer's waves balance is decreased by fee") {
-    val assetName = "myasset"
-    val assetDescription = "my asset description"
-    val f = for {
-
-      (firstAddressBalance, firstAddressEffectiveBalance) <- notMiner.accountBalance(firstAddress).zip(notMiner.accountEffectiveBalance(firstAddress))
-
-      issuedAssetId <- sender.issue(firstAddress, assetName, assetDescription, defaultQuantity, 2, reissuable = true, assetFee).map(_.id)
-
-      _ <- nodes.waitForHeightAraiseAndTxPresent(issuedAssetId)
-
-      _ <- notMiner.assertBalances(firstAddress, firstAddressBalance - assetFee, firstAddressEffectiveBalance - assetFee)
-        .zip(notMiner.assertAssetBalance(firstAddress, issuedAssetId, defaultQuantity))
-    } yield succeed
-
-    Await.result(f, waitCompletion)
-  }
-
-  test("Able to create asset with the same name") {
-    val assetName = "myasset"
-    val assetDescription = "my asset description"
-    val f = for {
-
-      (firstAddressBalance, firstAddressEffectiveBalance) <- notMiner.accountBalance(firstAddress).zip(notMiner.accountEffectiveBalance(firstAddress))
-
-      issuedAssetId <- sender.issue(firstAddress, assetName, assetDescription, defaultQuantity, 2, reissuable = false, assetFee).map(_.id)
-      _ <- nodes.waitForHeightAraiseAndTxPresent(issuedAssetId)
-
-      issuedAssetId <- sender.issue(firstAddress, assetName, assetDescription, defaultQuantity, 2, reissuable = true, assetFee).map(_.id)
-
-      _ <- nodes.waitForHeightAraiseAndTxPresent(issuedAssetId)
-
-      _ <- notMiner.assertAssetBalance(firstAddress, issuedAssetId, defaultQuantity)
-        .zip(notMiner.assertBalances(firstAddress, firstAddressBalance - 2 * assetFee, firstAddressEffectiveBalance - 2 * assetFee))
-    } yield succeed
-
-    Await.result(f, waitCompletion)
-  }
-
-  test("Not able to create asset when insufficient funds") {
-    val assetName = "myasset"
-    val assetDescription = "my asset description"
-    val f = for {
-
-      firstAddressEffectiveBalance <- notMiner.accountEffectiveBalance(firstAddress)
-      bigAssetFee = firstAddressEffectiveBalance + 1.waves
-
-      _ <- assertBadRequestAndMessage(sender.issue(firstAddress, assetName, assetDescription, defaultQuantity, 2, reissuable = false, bigAssetFee),
-        "negative waves balance")
-    } yield succeed
-
-    Await.result(f, waitCompletion)
-  }
-
-  val invalidAssetValue =
-    Table(("assetVal", "decimals", "message"),
-      (0l, 2, "negative amount"),
-      (1l, 9, "Too big sequences requested"),
-      (-1l, 1, "negative amount"),
-      (1l, -1, "Too big sequences requested")) //??? what message shoild be?
-
-  forAll(invalidAssetValue) { (assetVal: Long, decimals: Int, message: String) =>
-    test(s"Not able to create asset total token='$assetVal', decimals='$decimals' ") {
-      val assetName = "myasset2"
-      val assetDescription = "my asset description 2"
-      val decimalBytes: Byte = decimals.toByte
-      val f = for {
-        _ <- assertBadRequestAndMessage(sender.issue(firstAddress, assetName, assetDescription, assetVal, decimalBytes, reissuable = false, assetFee),
-          message)
-      } yield succeed
-
-      Await.result(f, waitCompletion)
-    }
-  }
-
-}
